# docs/contracts/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "Harness API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "projects"
                ],
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectPage"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "projects"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{projectId}": {
            "get": {
                "tags": [
                    "projects"
                ],
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "projects"
                ],
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "projects"
                ],
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{projectId}/status-digest": {
            "get": {
                "tags": [
                    "projects"
                ],
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "activityLimit",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectStatusDigestContract"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateProjectRequest": {
                "required": [
                    "organizationId",
                    "name",
                    "key",
                    "description"
                ],
                "type": "object",
                "properties": {
                    "organizationId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "criticality": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "repositoryUrl": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "repositoryProvider": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "defaultBranch": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "technologies": {
                        "type": [
                            "null",
                            "array"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "brand": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ProjectBrandContract"
                            }
                        ]
                    },
                    "memberProfileIds": {
                        "type": [
                            "null",
                            "array"
                        ],
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "additionalProperties": false
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "status": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "null",
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "detail": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "instance": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "ProjectPage": {
                "required": [
                    "items",
                    "nextCursor"
                ],
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectResponse"
                        }
                    },
                    "nextCursor": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "ProjectResponse": {
                "required": [
                    "id",
                    "organizationId",
                    "name",
                    "key",
                    "description",
                    "state",
                    "criticality",
                    "repositoryUrl",
                    "repositoryProvider",
                    "defaultBranch",
                    "technologies",
                    "brand",
                    "memberProfileIds",
                    "configVersion",
                    "chiefAgentId",
                    "operationMode",
                    "createdAt",
                    "lastActivityAt"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "organizationId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "criticality": {
                        "type": "string"
                    },
                    "repositoryUrl": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "repositoryProvider": {
                        "type": "string"
                    },
                    "defaultBranch": {
                        "type": "string"
                    },
                    "technologies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "brand": {
                        "$ref": "#/components/schemas/ProjectBrandContract"
                    },
                    "memberProfileIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "configVersion": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int64"
                    },
                    "chiefAgentId": {
                        "type": "string"
                    },
                    "operationMode": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastActivityAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProjectStatusDigestContract": {
                "required": [
                    "projectId",
                    "asOf",
                    "progress",
                    "taskCounts",
                    "attention",
                    "workflow",
                    "nextAction",
                    "recentActivity",
                    "unavailableSignals",
                    "fingerprint"
                ],
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "string"
                    },
                    "asOf": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "progress": {
                        "$ref": "#/components/schemas/ProjectProgressContract"
                    },
                    "taskCounts": {
                        "$ref": "#/components/schemas/ProjectTaskCountsContract"
                    },
                    "attention": {
                        "$ref": "#/components/schemas/ProjectAttentionContract"
                    },
                    "workflow": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ProjectWorkflowDigestContract"
                            }
                        ]
                    },
                    "nextAction": {
                        "type": "string"
                    },
                    "recentActivity": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectActivityContract"
                        }
                    },
                    "unavailableSignals": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "fingerprint": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
...
}